requirements.txt
decord
easydict
einops
gradio
numpy
pandas>=2.0.3
peft>=0.4.0
Pillow
tqdm
transformers==4.41.2
git+https://github.com/openai/CLIP.git
sentencepiece==0.2.0
protobuf
wandb
deepspeed==0.14.0
lmdb
msgpack
tensorboard
